Organize files and folders, add AddressForm and AssetDisplay
Files and folders are oraganized to follow general common practices.
During this process, Redux is applied to the project as well.

AddressForm and AssetDisplay  are temporarily named.
A better name will come near future with better structures.

There are few files(i.e. "App.css") that are hints traces of default
CreateReactApp format.
This will also be removed and modified in the future.