Implement multiple wallets
Improving this code to handle multiple wallets resulted in massive
modifications on Reducers and Dispatches - basically anything that
is related to state.
In the course of modifications, there were few uncomfortable code
and forced relationships that can be improved near future with further
refactoring and organization.

Lastly, this is the point where the missing tests are absoutely needed.